Refine summary serialization protocols
